Moved initial fetch logic to start page route
